+++ app/package.json
@@ -22,6 +22,7 @@
     "react": "^17.0.2",
     "react-color": "^2.19.3",
     "react-dom": "^17.0.2",
+    "react-intersection-observer": "^8.33.0",
     "react-lottie": "^1.2.3",
     "react-phone-number-input": "^3.1.25",
     "react-router-dom": "^5.2.0",
